+++ src/utils.ts
@@ -14,7 +14,7 @@ import { getConfig } from "./utils/pipeline/sfdxHardisConfig";
 
 export const RECOMMENDED_SFDX_CLI_VERSION = null; //"7.111.6";
 export const NODE_JS_MINIMUM_VERSION = 20.0;
-export const RECOMMENDED_MINIMAL_SFDX_HARDIS_VERSION: string = "6.1.4";
+export const RECOMMENDED_MINIMAL_SFDX_HARDIS_VERSION: string = "6.1.5";
 
 // Interface for execCommand and execSfdxJson options
 export interface ExecCommandOptions {